[MOOC] Added Sortable for Mooc elements
